scanner

+++ src/main/java/edu/bhscs/Clerk.java
@@ -1,17 +1,16 @@
 package edu.bhscs;
 
 public class Clerk {
-  public int ClerkCash = 0;
+  public Double ClerkCash = 0.0;
   public String cakePrice;
   Store myStore = new Store();
   Cake boughtCake = myStore.sellCake();
   Double cakeValue = myStore.getCakeValue();
 
-  public String cakeQualityMult(){
+  public String cakeQualityMult() {
     cakePrice = boughtCake.getQuality();
     return cakePrice;
   }
-
 
   public void speaken() {
     System.out.println("cake is for sale at the store currently");
@@ -20,7 +19,8 @@ public class Clerk {
             + boughtCake.getQuality()
             + " cake that weighs "
             + boughtCake.getWeight()
-            + " weight?, it costs " + cakeValue);
+            + " weight?, it costs "
+            + cakeValue);
   }
 
   public Double purchaseCake() {
